Moves combobox utils to separate file
